fix compiler warning

+++ Ipopt/contrib/RInterface/src/IpoptRJournal.cpp
@@ -31,7 +31,7 @@ void IpoptRJournal::PrintfImpl(Ipopt::EJournalCategory category,
 				  va_list ap) { 
     
     // Define string
-    const size_t MaxStrLen = 8192;
+    const int MaxStrLen = 8192;
     char s[ MaxStrLen ];
     
     // R guarantees to have an implementation of vsnprintf available
